WIP move changes done to domains out of it so that rest of the flows do not break
