+++ client/signup/steps/domains/index.jsx
@@ -46,10 +46,7 @@ import {
 	recordAddDomainButtonClickInTransferDomain,
 	recordAddDomainButtonClickInUseYourDomain,
 } from 'calypso/state/domains/actions';
-import {
-	getAvailableProductsList,
-	getMarketplaceProducts,
-} from 'calypso/state/products-list/selectors';
+import { getAvailableProductsList } from 'calypso/state/products-list/selectors';
 import getSitesItems from 'calypso/state/selectors/get-sites-items';
 import { fetchUsernameSuggestion } from 'calypso/state/signup/optional-dependencies/actions';
 import {
@@ -276,8 +273,6 @@ class DomainsStep extends Component {
 			  } )
 			: undefined;
 
-		const pluginItem = this.props.marketplacePlugin;
-
 		suggestion && this.props.submitDomainStepSelection( suggestion, this.getAnalyticsSection() );
 
 		maybeExcludeEmailsStep( {
@@ -296,7 +291,6 @@ class DomainsStep extends Component {
 				{
 					stepName: this.props.stepName,
 					domainItem,
-					pluginItem,
 					googleAppsCartItem,
 					isPurchasingItem,
 					siteUrl,
@@ -307,7 +301,6 @@ class DomainsStep extends Component {
 			),
 			Object.assign(
 				{ domainItem },
-				{ pluginItem },
 				this.isDependencyShouldHideFreePlanProvided() ? { shouldHideFreePlan } : {},
 				useThemeHeadstartItem
 			)
@@ -922,31 +915,13 @@ const submitDomainStepSelection = ( suggestion, section ) => {
 	);
 };
 
-function findMarketplacePlugin( state, pluginSlug, billingPeriod = '' ) {
-	const plugins = getMarketplaceProducts( state, pluginSlug );
-	const billingPeriodToTerm = {
-		MONTHLY: 'month',
-		YEARLY: 'year',
-	};
-	const term = ( billingPeriod && billingPeriodToTerm[ billingPeriod ] ) || '';
-
-	if ( ! term ) {
-		return plugins?.[ 0 ] || null;
-	}
-
-	return plugins?.find( ( plugin ) => plugin.product_term === term ) || null;
-}
-
 export default connect(
-	( state, { steps, flowName, queryObject } ) => {
+	( state, { steps, flowName } ) => {
 		const productsList = getAvailableProductsList( state );
 		const productsLoaded = ! isEmpty( productsList );
 		const isPlanStepSkipped = isPlanStepExistsAndSkipped( state );
 		const selectedSite = getSelectedSite( state );
 		const eligibleForProPlan = isEligibleForProPlan( state, selectedSite?.ID );
-		const marketplacePlugin = productsLoaded
-			? findMarketplacePlugin( state, queryObject?.pluginSlug, queryObject?.billingPeriod )
-			: null;
 
 		return {
 			designType: getDesignType( state ),
@@ -956,7 +931,6 @@ export default connect(
 			selectedSite,
 			sites: getSitesItems( state ),
 			userSiteCount: getCurrentUserSiteCount( state ),
-			marketplacePlugin,
 			isPlanSelectionAvailableLaterInFlow:
 				( ! isPlanStepSkipped && isPlanSelectionAvailableLaterInFlow( steps ) ) ||
 				[ 'pro', 'starter' ].includes( flowName ),
